Comment out the ESO installation via subchart
This way it won't fight the ESO installed via the community operator
but we still get the clustersecretstore and the service account that
are used in the imperative namespace to unseal, init and configure
the vault.

This is a short-term approach while we evaluate our more mid/long term
options.

I have tested the multiple ClusterSecretStore object. I did not
test that commenting out the dependencies in Chart.yaml, hopefully
that just works :)